Make `Layout.template_name` a unique field.
There should be no reason to have several layouts that use the same template
under a different name for different content types. If a template applies to
multiple content types, just add all of them to the one layout and give it a
name that describes the template, not the content types it applies to.