Fixed raw_id fields pointing to polymorphic models.
Previously, foreign keys pointing to "child" polymorphic models would
not work as raw id fields in the admin, throwing 400 errors (due to
invalid _to_field parameters in the URL). This fix replaces the widgets
for raw id fields in both fluent content pages and the plugin inlines
with ones that use the correct URL for the pop-up.

As a bonus, this resolves issue #144 raised by @tailee in django-
polymorphic:
https://github.com/chrisglass/django_polymorphic/issues/144