Refactor publishing to remove monkey-patch hackery approach, re #38
Refactor the base publishing model and the publishing app's AppConfig
to replace the post-hoc monkey-patching approach with a standard, and
more comprehensible & manageable, normal model/admin inheritance
approach.

This change is made on the assumption we don't want or need to do
post-hoc monkey-patching of models in the AppConfig, and will instead
make publishing-compatible models available as part of ICEKit (e.g.
PublishableFluentContentsPage etc.)