Enable/disable ICEKit publishing features in admin, re #38
Enable or disable publishing features of ICEKit models in the CMS admin
depending on whether or not the 'icekit.publishing' app is installed in
the project.

This is a rudimentary mechanism for configuring publishing support,
which will not prevent most of the publishing-related work from being
done such as adding extra model fields for publishing status etc.