Rename all "Publisher*" items to "Publishing*", re #5
To improve the naming of models, fields, etc in ICEKit's publishing
implementation we will ditch the "Publisher" convention used by the
django-model-publisher library -- on which we are no longer dependent --
and instead switch to using "Publishing" as the default naming prefix.

This will also let us clearly distinguish between objects that use the
legacy/original publishing implementation, and those based on ICEKit's
implementation.