Add signals to handle saving of draft instance on (un)publish, re #5 #23
Add new signals `publishing_publish_save_draft` and
`publishing_unpublish_save_draft` that do a very simple job by default
-- save a draft instance object -- but which can be disconnected and
replaced in downstream code to do more complicated things such as save
versioning data.

It is a bit cumbersome and slow to use signals for this job instead of
subclass overriding, but it is much more flexible and easier to override
in downstream projects rather than needing to insert

NOTE: This might all be made unnecessary if we add full support for
versioning (with reversion) to ICEKit later.