Make field tracker respect `MONITOR_FIELDS` overrides, see SFMOMA #1576
Add `FieldTracker` to event classes dynamically rather than in class
definition so subclass changes to `MONITOR_FIELDS` are respected.

The `MONITOR_FIELDS` class attribute is intended to be overrideable in
subclasses so we can track/monitor additional fields, but this did not
work correctly because – when applied in the standard way – the
`FieldTracker` utility we use can only track fields present in the
exact class on which it is defined, not fields added in subclasses.

This meant we would need to redefine the `tracker` class attribute in
any subclasses that added monitored fields, otherwise changes to these
fields would not be tracked at all. Overriding the tracker in this way
would be easily forgotten, error-prone, and awkward in code, so to try
and prevent regular bugs in downstream projects I opted to attach the
tracker dynamically instead of at class preparation time.

This issue was somewhat hidden because the field propagation logic based
on `MONITOR_FIELDS` and applied via `get_repeat_fields()` worked in all
cases, it was only detection of field changes via the `tracker` that
would break. And this would only be detectable if subclass-only
monitored fields were changed without any changes at the same time to
fields on `AbstractEvent`.