Replace field tracker hackery with sanity-checkingfor devs, see SFMOMA #1576
Instead of forcibly hacking the field tracker to respect customised
`MONITOR_FIELDS` listings in event subclasses – an approach which is a
little ugly and might not be robust – we will instead raise an exception
if developers write an event subclass with monitored field changes that
are not properly tracked. It will then be up to the developer to fix the
issue by redeclaring an appropriate field tracker attribute in their
subclass.

This will allow developers to see these errors early on, and will
provide clear instructions to fix the problem.