Fix excessive & unnecessary DB queries when querying published items
Check a model FK attribute to detect whether an an item has a related
published copy, instead of loading and throwing away the entire
published copy object.

This was a particularly brutal error because the `has_been_published()`
method is called many times in the codepath for the frequently-used
`published()` filter.