Use more efficient DB queries when possible for publishing exchange
When exchanging draft-for-published copies use more efficient DB
queries when possible. And only perform the exchange query and
re-ordering work if it is really necessary.

Also be more defensive when checking published status for items when
the queryset is not for a `PublishingModel` model, since there are no
guarantees for which fields will be available.