Fix delete collector for 1.8 to avoid DB integrity errors
Our patched version of the Django 1.8 delete collector was not finding
all the parents of items in all cases, which meant it would not find
and handle deletes across relationships to the `HtmlPage` proxy model
leading to errors like the one below.

This fixes the lookup of parent models to find them all.

    IntegrityError: update or delete on table "fluent_pages_urlnode"
    violates foreign key constraint
    "fluent_pa_master_id_5300b55ee85000a1_fk_fluent_pages_urlnode_id"
    on table "fluent_pages_htmlpage_translation"
    DETAIL:  Key (id)=(26933) is still referenced from table
    "fluent_pages_htmlpage_translation".