Fix finicky requirements to allow installation for Django 1.7 project
`django-el-pagination` versions 3+ drops support for Django < 1.8 and
this directive from that library overrides anything we can say in our
setup.py or Tox configs to really install Django 1.7, and instead
installs Django 1.10 (or whichever version is latest) instead.

The '<3' version pinning *must* be done in the top-level
`install_requires` because doing it later in the "django17"
`extras_require` does not properly override the version, and we still
end up getting django-el-pagination>=3 and therefore Django 1.10.