Respect ICEKIT_DIR environment variable when finding ICEKit install path
For projects based on ICEKit the path to the ICEKit directory within
a Docker container isn't necessarily under 'icekit' in the base
directory. In these situations we already have an environment variable
that points to the correct location, so this change uses it.

This change is motivated by a downstream project in which the Node/NPM
installation of ICEKit requirements was performed in an /opt/ location
but the project then looked for the `lessc` utility in ICEKit's venv
installation path where it was not available.