Run `entrypoint.sh` via `ENTRYPOINT` instruction in `Dockerfile`.
It's only configuring the environment now, not actually running or
connecting to anything, so it should be much faster and safer than it
has been in the past.

Previously, we moved it from `ENTRYPOINT` to `CMD` because it was
unstable and a PITA to workaround/disable when encountering a fatal
error.

This should simplify most run commands and reduce the probability of
accidentally running another script that requires entrypoint.sh to have
been run already, without having run it.