Add requires.io badge.
It's OK to be "outdated" if we are intentionally pinning some packages to older versions (e.g. Django LTS). If it ever changes to "insecure", we need to unpin and deal with any breaking changes.

A pull request will be created by requires.io automatically, if it does ever change to "insecure".