Make publishing templates work for objects without absolute URLs
The publishing templates did inline lookups of `get_absolute_url()` on
objects to get preview URLs, but this failed for objects which are not
URL routable (such as Events in at least some situations).

The `get_absolute_url()` lookups are now done in the admin instead,
with exception handling to avoid errors for unroutable objects. These
errors likely only showed in development mode, but made admin pages
unusable in this case.

See ICEKit ticket #154 in Assembla