Use `local` instead of `latest` tag for Docker images.
The `latest` tag is badly named. We never actually push a `latest` tag
to Docker Hub. We only push release tags and branches (0.1.2, develop,
master, etc.)

The default `Dockerfile` and `docker-compose.yml` files are geared
towards building and running a local image, so let's call it that.

The `startproject.sh` script will update `Dockerfile` and
`requirements-icekit.txt` to use release versions of ICEkit.