Add SlugMixin to centralise slug uniqueness checks
The new `SlugMixin` abstract model class adds a `slug` field to models,
but more importantly also checks somewhat intelligently that the slug
is unique when the item is saved, or when published if the item is also
publishable.

As part of this change applied a counter-intuitive, and possibly wrong-
headed, fix to ensure the `publishing_prepare_published_copy` method
that can be implemented by subclasses of `PublishingModel` actually
gets called by treating it as an ancestor method.

See ICEKit ticket #154 in Assembla