Move polymorphic child model admin fix from FluentLayoutsMixin, re #31
The `PolymorphicChildModelAdminGetParentAdminFix` previously included
in `FluentLayoutsMixin` is not appropriate there because not all
implementors of the `FluentLayoutsMixin` admin are for polymorphic child
models.

Unfortunately I can't find a better central location to hook this work-
around into the system without triggering import errors. For the time
being the best approach is to add the mixin admin fix
`PolymorphicChildModelAdminGetParentAdminFix` to every fluent page
model, which is pretty terrible.

Longer-term we need to find a better place or way to do this.