Merge `AbstractEvent` and `Event` and rename to `EventBase`
Simplify models into a single concrete model named `EventBase` which is
the same as the prior `Event` model, but merged with the now defunct
`AbstractEvent`.

This seemingly simple change had cascading knock-on effects and required
a hacky DB migration to avoid Django getting confused and prompting to
remove stale content types that it should have removed itself.