Fix generation of all-day events to include repeat end date
The prior fix to RRULE generation of repeat all-day events to include
the repeat end date did not apply to the `generate()` generator method
which uses the `between()` method to constrain generated datetimes. So
we apply the same kind of fix in a second place with subtle differences
due to the different inclusion/exclusion rules between RRULE specs and
the python `rruleset.between()` method.