Work around sed working differently on Mac and Linux
On Linux, sed requires either an "=" or no space after the "-i" argument, while Mac expects a space, but can handle there not being one, except if an empty extension is desired, which is what we want. Thus we have to resort to specifying an extension, then deleting the backup files. See discussion here: http://stackoverflow.com/questions/5694228/sed-in-place-flag-that-works-both-on-mac-bsd-and-linux