Require integer IDs in objects relatable to workflow steps, re #156
For the `object_id` Generic Foreign Key field we were using a
`CharField` to allow for target models that don't have an integer
primary key ID, which would be an unusual case but is possible.

However this extra flexibility breaks querying over generic
relationships in Postgres 9+ due to the type mismatch between the
actual integer IDs and the character value of `object_id`. Such querying
should work but doesn't, see the unresolved bug report
https://code.djangoproject.com/ticket/16055

Supporting non-integer PKs on target objects would be nice, but not nice
enough to warrant the extra work this would require right now (if it is
possible at all).