Improve workflow status filtering implementation, re #156
Add new full class implementations for filtering by workflow state
`status` and `assigned_to` fields.

With these classes we can apply the filters to models that do not
provide a generic relation and model attribute named `workflow_states`,
which is important for page models in particular where we want to avoid
monkey-patching in fields like this if we can avoid it.

If a model does provide an appropriate generic relation by extending
`WorkflowStateMixin` then the filter uses the reverse relationship to
do its work, and the code path is much simpler.

If the model does not provide the reverse relationship we do more work
to filter the model instances appropriately.