Add method to avoid multiple model registrations on ICEkitURLField
The new class method `ICEkitURLField.register_model_once` only registers
a given model with `ICEkitURLField` if it is not already registered.

This gives us a convenient way to register models in application ready
methods without the risk that the model will get re-registered (and
fail with an error) when the app is run in unit testing conditions.

This change was prompted by the model registration in ICEkit Events
being called twice during unit test runs, causing the tests to fail.