Avoid booby-trapping publishing_draft model relationship
The `publishing_draft` reverse relationship attribute for publishable
items should always return the raw draft copy, never a draft copy
wrapped in `DraftItemBoobyTrap`.

For some reason I don't understand the `publishing_draft` reverse
relationship attribute on polymorphic models could return a
booby-trapped draft copy instead of the raw copy in some cases.

With this change these the `publishing_draft` relationship is not longer
mistakenly wrapped in this case.

This issue first surfaced when it affected the `WorkBase` model from
glamkit-collections, but strangely we have not seen it before in other
polymorphic models like fluent pages where we had tests for the
correctness of draft traversal.