Fix workflow status & assigned admin filters for polymorphic models
We cannot use simple queryset filtering for workflow status or assigned
values when we are working with polymorphic models – even if they
extend `WorkflowStateMixin` – because we will probably get a queryset of
parent items which is not aware of the GFK reverse relationships to
`WorkflowState`.

Solve this problem using brute force for now, by bypassing the basic
filtering approach for polymorphic querysets and using the slower
concrete instance iteration process instead.