Remove `publishing_draft` relationship hack
The monkey-patch intended to prevent `publishing_draft` from being
wrapped by `DraftItemBoobyTrap` was not working broadly enough for
icekit-events, and seems to be fragile and broken in general, so I am
removing it for now.

We will need to fall back to always using `get_draft()` to be sure we
get the unwrapped draft item, and put up with the unnecessary wrap-then-
unwrap overhead.

Revisit this if/when we have time.