Fix bug preventing publishing of anything with link item content items
Remove the overly magical behaviour where link items like `PageLink`
and `ArticleLink` etc automatically delegated attribute calls to the
target linked item for all but two explicit attribute names.

This automatic delegation is fundamentally a bad idea, because it makes
attribute & method usage on one object do work on another object
altogether.

The specific reason for removing this now is because we were unable to
publish any content with link content items in ACMI because the
publishing system did work on the `publishing_linked` relationship of
the link content item, and ended up getting the published copy of the
target linked item instead.

Rather than working around the issue by adding `publishing_linked` as
another special-case magical attribute name that is handled sensibly,
and leaving this bomb to explode again later, I am removing the
behaviour altogether.

WARNING: This may break some page rendering or other situations where
the magical behaviour was used to do work on the linked target instead
of the actual link content item that is present.