Fix odd DB migration behaviour for workflow app's model, re #190
Explicitly set `app_label` for the `WorkflowState` model to ensure
Django treats it as belonging to the "icekit_workflow" app, instead of
treating it as belonging to "icekit" instead.

This should fix a situation seen in AOT where running the makemigrations
command produced two bogus DB migrations, one in the
icekit/migrations folder to create a WorkflowState model in the wrong
place, and one in icekit/workflow/migrations to remove the model from
the right place.