Add migration operation to help rename portable applications
Add the `RenameAppInMigrationsTable` operation which helps to handle
the difficult situation where a portable application has been renamed
but migrations with the old name have already been applied in a project.

Previously we handled this situation by requiring that deployers
remember to manually run a SQL command in the target DB *before* running
the updated DB migrations. That SQL would look something like:

    UPDATE django_migrations SET app='newname' WHERE app='oldname';

The new `RenameAppInMigrationsTable` operation automates running of this
UPDATE command, if it is necessary.

It does not solve the problem where the migrate command must be re-run
to detect the changed history in the `django_migrations` table, but it
makes it clearer what is going on with a descriptive error message that
tells the user what has been fixed and to re-run the migrate.

NOTE: This new operation has only been tested with and applied to the
'workflow' => 'icekit_workflow' app rename so far, but we should use
it in all cases in the future.