Fix cropping in IIIF Image API #209 #204
Fix crop (region) and resize (size) image transformation in IIIF API
to work as expected.

To do this we ditch the easy-thumbnails default transform operations,
which do not work anything like what we need, and instead do more
direct and explicit manipulations using the PIL Image API.

Taking on more control entails more work to load, save, and
colour-convert the image, but this change is worth it (and required)
in the long run.

NOTE: Since we no longer use easy-thumbnails standard image load/save
features we have lost the ability, for now, to load/save transformed
images from/to a storage system.