#213 Update tests to use "api" URL reverse namespace
Adding the `namespace` parameter to the API's URL
definitions -- to allow us to set the `app_name`
so we can easily recognise API calls and let them
through the publishing checks -- did have knock-on
side-effects in that now all URL reverse calls
must use the "api:" namespace prefix.

Fix unit tests to use this namespace, and hope we
don't get bitten more by the over-specified URL
patterns in the future.