#213 Make API available at api.HOSTNAME domain name
Use `django-hosts` library to make the API endpoints
available at the api.HOSTNAME sub-domain name, instead
of at the HOSTNAME/api/ URL path.

NOTE: The HOSTS base settings section had to be
moved to the bottom, below whitenoise etc in particular,
to ensure the `HostsResponseMiddleware` item was at the
end of the MIDDLEWARE_CLASSES config.