#213 Add easy write for API serializers
The `WritableSerializerHelperMixin` helper/mixin class is
now available to handle writable related fields in a general
way in a serializer, without requiring custom code in every
case where a writable item has writable relationships.

The `WritableSerializerHelperMixin` is a renamed version of
the prior `ModelSubSerializerParentMixin` and still handles
`ModelSubSerializer` fields transparently, no extra work
required.

For other related fields based on the standard
`ModelSerializer` you can now add field-specific settings
as `Meta.writable_related_fields` to map field names to
`WritableRelatedFieldSettings` objects that define how to
look up a related model instance, and whether to permit
write operations to create or update related instances.

NOTE: This commit adds a dependency on the `attrs` library
to make definition of a small class easier, and because I
want to use it in general.

NOTE 2: This is a first pass at handling a complex set of
situations, which should really be handled in the DRF base
library but do not seem to be. Consider this implementation
to be experimental for now, and it is definitely incomplete.