#213 Un-pin DRF filters so DRF version 3.6.3 will work
The latest djangorestframework version 3.6.3 did not work
properly with the 0.9 version of djangorestframework-filters
we had pinned.

So -- following a fix in glamkit-collections to work with
the newest DRF filters -- this change undoes the pinning to
install the latest DRF and DRF filters versions which do
actually work with each other.