#236 Add `brief` and `admin_notes` fields to models with workflow
Add `brief` and `admin_notes` text fields to the
`WorkflowStateMixin` abstract model used as a basis
for models that maintain workflow relationships.

This gives us a central place to store general
information documents about items that are
(probably) useful in a workflow context.

These fields are so far exposed only in item
listing admin pages.