Extra DB migrations following upstream change in ICEkit
The `WorkflowStateMixin` model in django-icekit -- which
is used as a basis for the `EventListingPage` model -- was
updated with two new fields: `brief`, and `admin_notes`.

This change updates the model in this project to comply
with the upstream changes.

And will hopefully make ICEkit unit tests pass again in Travis.