#213 Improve model sub-serializer's field name mangling
Change the approach `ModelSubSerializer` uses to rename
fields to group and strip the prefix – e.g. `name_full`
to `name.full` – so that it works better and will
report the correct field names to external interfaces.

Also bump glamkit-collections version which has a
related fix in usage of the sub-serializer.