Fix key error in publishing admin's handling of unique fields
The publishing admin does some special handling to
clean unique field values submitted in the form.

If a valid value was not provided for a form field
specified as unique, the admin would crash with a
`KeyError` doing this custom cleaning work.

E.g. when submitting an admin form for a titled
model with required slug value, where the slug is
also a unique field, if no slug or an invalid slug
was provided the submit would cause a 500 error.