Add nearby location constraints to advanced event listing page
The advanced event listing page now accepts the
'is-nearby' GET parameter with comma-delimted
latitude/longitude/distance values and will
restrict event results to those with locations
within the given distance of the visitor's (or
given) location.

The default template for this page prompts the
visitor to permit location lookup to populate
the lat/longs, or they can be set manually.