Subtle fix to make publishing `get_actions()` method work
The `PublishingAdmin` class was sometimes not
correctly inheriting the `get_actions()` method
from `_PublishingHelpersMixin`: it would be called
as expected for the Pages parent admin, but
not for the Events or Location parent admins.

Swapping the ordering of base classes inherited
by `PublishingAdmin` somehow fixes this oddity,
for reasons I cannot fathom.