Pin DRF dependencies to work with Django 1.8
Pin dependencies `djangorestframework<3.7` and
`django-filter<1.1` which are versions that work
with Django 1.8. Later versions produce mysterious
errors.

I had to apply these version constraints in a
nasty global way because I could not find a way
to specify the versions in setup.py that would
actually be respected by setuptools. And I really
tried.

I also had to apply the pinning in two places:
- setup.py for general usage and the go.sh
  environment
- requirements.txt for Travis CI builds.