Modify setup-django.sh to use yarn instead of bower and npm if available re #286,
currently setup-django makes the assumption that yarn.lock exists which it may not initially.

In order to use BS4 and compile from source, we need to move away from bower (at least to install bootstrap, since bower is deprecated), they recommend using yarn instead. Can we get away with simply using npm and django-compressor?

The goal is to have:

 - bootstrap 3 available to the admin
 - bootstrap 4 available to the front end
 - icekit .scss files compiled into the project/overridable where necessary
 - project .scss files that allow for bootstrap variables and overrides and provide a platform to build on that simplifies frontend setups for icekit sites.
 - can we get away with using only django-compressor or is yarn/alternative a necessary step?
 - as simple a setup as possible to avoid painful edge cases and implementation nightmares
 - sourcemaps would be nice

Yarn

 - use bower-away to convert bower.json to yarn (adds yarn specific syntax to package.json)
 - yarn requires Ruby > 2.3
 - you might need to delete bower_components/.DS_Store if bower-away fails with ENOTDIR