Update publishing of M2Ms to work in more situations
When publishable items are processed there is a
lot of special handling required to keep many-to-
many (M2M) relationships in sync between draft and
published copies.

This change makes this handling work with M2M
relationships that have a through table model with
`GenericForeignKey` (GFK) fields pointing to the
source and/or target of the M2M relationship.

The main motivation is to make ICEKit Publishing
compatible with the django-taggit library which
uses a GFK to the source object in its `TaggedItem`
through model. But this change should handle all
similar cases.

NOTE: To work with django-taggit and in
potentially other cases the M2M manager associated
with publishable items must provide the
`source_field_name` and `target_field_name` field
name attributes usually included in M2M managers.
If the manager does not already have these fields,
it must be modified to have them.

See https://github.com/ixc/biennale-of-sydney/issues/23

TODO: Build out unit tests to include publishing
of M2M-with-through-model relationships that use
GFKs for source and target fields.