#308 Don't load page type select form when adding LayoutPage
Avoid test failures loading the page type selection
form when adding layout pages, and instead load
the Add page for the specific `LayoutPage` content
type by adding the `ct_id` GET parameter.

This change was necessary to fix the tests for
django-fluent-pages version 1.1.3.

WARNING: I am not sure that this change is hiding
a real issue present in Fluent Pages 1.1.3 that
should be fixed properly, rather than doing what
this change does which is just avoid triggering it.