Fix auto-publishing to include model relationships
Only trigger the automatic publish action when
publishable instances are saved in the admin,
and *after* all that instance's relationships
have also been finalised and saved.

Previously, when a new publishable instance was
saved it was immediately published with the default
`post_save` signal, but this could occur before the
item's relationships had been properly created or
updated.

This surfaced in BOS as a fault where a Fluent
Page's related translations were not created before
a newly-created item was auto-published, leading
to orphaned bogus translation records and missing
translations in the published copy of the newly-
created page.

But in fact, problems with any model with complex
relationships where inevitable with the old
approach.

NOTE: With this change, auto-publishing will only
work via the CMS admin pages, not for code-level
model saves.

See https://github.com/ixc/biennale-of-sydney/issues/45