Do not try to fetch previews for AnyURLFields
AnyURLFields result in multiple input fields with "_#" appended, some of
which may be foreign key widgets. But even if we detect those and remove
the prefix, we are then still unable to look up the model that's being
referenced, since it's still just a CharField, not a ForeignKey. To be
able to do that, we would need a lot of anyurl-specific code.

In the meantime, simply avoid requesting previews for AnyURLFields,
which result in fewer 500 errors cluttering the logs, since that's what
such requests will generate.